requirements.txt
streamlit==1.31.0
sendgrid==6.11.0
python-dotenv==1.0.1
fastapi==0.115.0
pydantic==2.9.2
uvicorn==0.32.0